added util.memory
