+++ distribution/all/store.js
@@ -1,0 +1,3 @@
+const memory = require('../util/memory');
+
+module.exports = memory('store');
